Refactor imports in various files to use TableInterfaces for consistency and clarity
