+++ src/interfaces/application/Event.ts
@@ -1,12 +1,11 @@
 import { User } from "../domain/User";
 import { Component, BaseSelectMenu } from "./Message";
 import { Duration } from "../../utils/Duration";
-import { ServersModel } from "../database/TableInterfaces";
+import { ServersModel, TimelineEntriesSaveModel } from "../database/TableInterfaces";
 import { Interaction as DiscordInteraction, Message as DiscordMessage } from "discord.js";
 import { MultiLingualString } from "../../utils/i18n/MultiLangualString";
 import { Command } from "./Command";
 import { Games_Settings, GameSettingsSchema, GameSettingsValues } from "../domain/GameSettings";
-import { TimelineEntriesSaveModel } from "../database";
 
 export enum EventTypeEnum {
     SLASH_COMMAND = "SLASH_COMMAND",
